Implement main model class and add folder for database adapter
Implement the database class. It will be the top class for accessing to
database datas.
This class need a adapter to store data y using a specific technology
All adapter will be in adapter/ folder
The init file of this folder contains the base class and some utilies functions
to help you to write your own database adapter